sysdb logic -> DAOs
